--- .changeset/dry-balloons-smell.md
@@ -1,5 +1,0 @@
----
-"@smithy/protocol-http": minor
----
-
-add RequestHandlerParams type for ctor passthrough
