+++ .changeset/sixty-fireants-pay.md
@@ -2,6 +2,7 @@
 "@smithy/fetch-http-handler": minor
 "@smithy/node-http-handler": minor
 "@smithy/protocol-http": minor
+"@smithy/types": minor
 ---
 
-add static factory for http handlers
+move default fetch and http handler ctor types to the types package
